controller added for easy code management
