+++ backend/routes/auth.js
@@ -7,68 +7,9 @@ const router = express.Router();
 
 require("dotenv").config();
 
-router.post("/register", async (req, res) => {
-  const { firstName, lastName, email, password, roleType } = req.body;
-  try {
-    let user = await User.findOne({ email });
-    if (user) {
-      return res.status(400).json({ msg: "User already exists" });
-    }
-    user = new User({
-      firstName,
-      lastName,
-      email,
-      password,
-      roleType,
-    });
-    await user.save();
-    const payload = {
-      user: {
-        id: user._id,
-      },
-    };
-    const token = jwt.sign(payload, process.env.JWT_SECRET, {
-      expiresIn: "1h",
-    });
-    res.status(201).json({ token });
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send(`SERVER ERRROR ${err.message}`);
-  }
-});
+router.post("/register", register);
 
-router.get("/register", async (req, res) => {
-  res.status(500).send(`SERVER CANNOT RESPOND ${req.message}`);
-});
 
-router.get("/login", async (req, res) => {
-  res.status(500).send(`SERVER CANNOT RESPOND ${req.message}`);
-});
-
-router.post("/login", async (req, res) => {
-  const { email, password } = req.body;
-  try {
-    const user = await User.findOne({ email });
-    if (!user) {
-      return res.status(400).json({ msg: "Invalid credentials" });
-    }
-    const isMatch = await user.matchPassword(password);
-    if (!isMatch) {
-      return res.status(400).json({ msg: "Invalid credentials" });
-    }
-    const payload = {
-      user: {
-        id: user._id,
-      },
-    };
-    const token = jwt.sign(payload, process.env.JWT_SECRET, {
-      expiresIn: "1h",
-    });
-    res.json({ token });
-  } catch (err) {
-    console.error(err.message);
-    res.status(500).send("SERVER ERROR");
-  }
-});
+router.post("/login",login);
 
 module.exports = router;
